A few clang-tidy suggested fixes (missing returns/cases, const refs where safe ect.)

+++ src/parser/inc/MatrixNode.hpp
@@ -14,7 +14,7 @@ class MatrixNode : public ExprNode {
    public:
     std::vector<std::vector<std::shared_ptr<ExprNode>>> data;
 
-    int numElements();
+    size_t numElements();
     std::vector<int> getDimensions();
     llvm::APInt genAPIntInstance(int numElements);
     llvm::Type* getLLVMType(llvm::Module* module);
